Use exact FP32 W8A8 test reference on CUDA

+++ tests/kernels/test_w8a8_linear.py
@@ -19,6 +19,9 @@ def test_w8a8_linear_matches_per_row_quantization_reference() -> None:
     qweight, wscale = quantize_weight_per_channel(weight)
     scale = x.float().abs().amax(dim=1, keepdim=True).clamp_min(1e-8) / 127
     qx = torch.round(x.float() / scale).clamp(-127, 127).to(torch.int8)
-    reference = (qx.to(torch.int32) @ qweight.t().to(torch.int32)).float() * scale * wscale.float()[None, :]
+    # CUDA does not implement int32 @ int32 in PyTorch. At K=1024 the largest
+    # signed INT8 accumulation is 16,516,096 (< 2**24), so FP32 represents this
+    # integer reference exactly on the T4.
+    reference = (qx.float() @ qweight.t().float()) * scale * wscale.float()[None, :]
     actual = w8a8_linear(x, qweight, wscale)
     torch.testing.assert_close(actual, reference.to(torch.float16), rtol=4e-3, atol=4e-3)
